Add a failing test for undef params
I've been using this pattern whereby the parent class proxies params to the
relevant sub-classes to provide a single entry point, whilst keeping the
defaults and validation as close to the classes that use them.

However this brings up an issue with Puppet 3.x whereby the `undef` defaults
in the parent class appear to be cast as empty strings. Meaning that
`include openconnect` doesn't raise a warning and the templates get rendered
with empty strings.

Adding this test here as a reminded, until we figure out the desired
solution.