Change default and logic for dnsupdate
This matches the upstream default of true (`yes`). The `vpnc-script` only
checks for a value like `/no/i`, so we should omit the environment variable
if the param is given a `true` boolean.

I've used `if @dnsupdate == false` instead of `unless @dnsupdate` to make it
clearer what we're matching. The selector has been removed since we're only
interested in a `no` value.