Another attempt at fixing dep versions
It seems that the `Error: undefined method `MIN' for SemVer:Class` error
isn't caused by simple whitespace in the version field. My feeling is that
it probably expects a three-point version number.

What we're trying to fix/workaround is the `_warnings` field being emitted
by the following request, which chokes the `puppet module` tool:

➜  openconnect git:(master) ✗ curl http://forge.puppetlabs.com/api/v1/releases.json\?module\=gdsoperations/openconnect\&version\=0.0.4
{"gdsoperations/openconnect":[{"file":"/system/releases/g/gdsoperations/gdsoperations-openconnect-0.0.4.tar.gz","version":"0.0.4","dependencies":[["puppetlabs/stdlib",">=3.0"]]}],"puppetlabs/stdlib":[],"_warnings":["Dependency puppetlabs/stdlib for module gdsoperations/openconnect had invalid version range \">=3.0\""]}%

It's a shame that the only feedback loop for this is to cut a new release,
upload it to the Forge, and then attempt to download it :(